Added Code File
This project was developed in late 2024.

This is a face detection attendance system/ mock employee monitoring software assuming that employees are working in an online environment where their cameras will be turned on
By determining the attendance based on detected faces, it optimizes the attendance-taking process and determines engagement through the following test cases:
Have all the employees attended? 
Have none of the employees attended? 
How many employees are absent? 
How many extra attendees (possibly hacked meeting) are present? 
This program also has other features like being able to login and quit the program, 
printing to the frame for detected attendees, recording and saving videos, adding 
timestamps to the file names of these videos for the host to have proof of attendance, 
the intuitive terminal that seeks for user inputs, and engaging sound effects.